Get demo background functions working with Inngest
