+++ package.json
@@ -20,6 +20,7 @@
     "eslint": "8.22.0",
     "eslint-config-next": "13.1.1",
     "fastest-levenshtein": "^1.0.16",
+    "inngest": "^0.9.3",
     "lodash": "^4.17.21",
     "next": "13.1.1",
     "react": "18.2.0",
