+++ pages/api/inngest.ts
@@ -1,0 +1,4 @@
+import { serve } from 'inngest/next';
+import addToCatalog from './cron/inngest/functions';
+
+export default serve('My app name', [addToCatalog]);
